@@ -9,9 +9,10 @@
   ],
   "scripts": {
     "start": "gulp server",
+    "demo": "SINGLE_DEMO=true gulp server",
     "build": "gulp clean && gulp sass && gulp copy && webpack",
     "prepublish": "babel ./src --out-dir ./lib && gulp dist",
-    "test": "eslint src && jest",
+    "test": "npm run lint && jest",
     "test:watch": "jest --watch",
     "lint": "eslint src",
     "gen": "node examples/scripts/generateExampleConfigs.js && node examples/scripts/generateExamples.js && xdg-open docs/jquery.html",
@@ -32,20 +33,19 @@
     "react-component"
   ],
   "devDependencies": {
-    "prettier": "^1.14.3",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
     "autoprefixer": "^7.1.2",
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.16.0",
-    "babel-eslint": "^7.0.0",
-    "babel-jest": "^19.0.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-assign": "^6.8.0",
-    "babel-polyfill": "^6.16.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^9.0.0",
+    "babel-jest": "^24.8.0",
+    "babel-loader": "^8.0.4",
     "babel-preset-airbnb": "^2.1.1",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "css-loader": "^0.28.0",
+    "css-loader": "^2.1.1",
     "deepmerge": "^1.1.0",
     "del": "^2.2.2",
     "enzyme": "^3.2.0",
@@ -56,29 +56,32 @@
     "eslint-plugin-react": "^6.3.0",
     "express": "^4.14.0",
     "foundation-apps": "^1.2.0",
-    "gulp": "^3.9.1",
-    "gulp-sass": "^3.1.0",
+    "gulp": "^4.0.0",
+    "gulp-sass": "^4.0.0",
     "husky": "^0.14.3",
     "jasmine-core": "^2.5.2",
-    "jest": "^19.0.2",
+    "jest": "^24.8.0",
     "jquery": "^3.2.1",
     "js-beautify": "^1.7.5",
     "json-loader": "^0.5.4",
     "lint-staged": "^7.0.3",
     "node-sass": "^4.5.2",
+    "opn": "^5.4.0",
     "postcss-loader": "^1.3.3",
+    "prettier": "^1.14.3",
     "raf": "^3.4.0",
     "react": "^16.0.0",
     "react-addons-test-utils": "^15.6.2",
     "react-dom": "^16.0.0",
     "react-test-renderer": "^16.0.0",
-    "run-sequence": "^1.2.2",
     "sass-loader": "^6.0.3",
     "sinon": "^2.1.0",
     "slick-carousel": "^1.8.1",
     "style-loader": "^0.16.1",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.16.1",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "webpack": "^4.21.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.9",
     "why-did-you-update": "^0.1.1"
   },
   "dependencies": {
