@@ -3,8 +3,8 @@
 import React from "react";
 import classnames from "classnames";
 
-var getDotCount = function(spec) {
-  var dots;
+const getDotCount = (spec) => {
+  let dots;
 
   if (spec.infinite) {
     dots = Math.ceil(spec.slideCount / spec.slidesToScroll);
@@ -17,7 +17,7 @@ var getDotCount = function(spec) {
   return dots;
 };
 
-var getNthDotAssociatedSlides = function(n, spec) {
+var getNthDotAssociatedSlides = function (n, spec) {
   var nthDotAssociatedSlides = "";
   for (
     var i = spec.slidesToScroll * n;
@@ -44,11 +44,11 @@ export class Dots extends React.PureComponent {
     this.props.clickHandler(options);
   }
   render() {
-    var dotCount = getDotCount({
+    let dotCount = getDotCount({
       slideCount: this.props.slideCount,
       slidesToScroll: this.props.slidesToScroll,
       slidesToShow: this.props.slidesToShow,
-      infinite: this.props.infinite
+      infinite: this.props.infinite,
     });
 
     // Apply join & split to Array to pre-fill it for IE8
@@ -56,29 +56,30 @@ export class Dots extends React.PureComponent {
     // Credit: http://stackoverflow.com/a/13735425/1849458
     const { onMouseEnter, onMouseOver, onMouseLeave } = this.props;
     const mouseEvents = { onMouseEnter, onMouseOver, onMouseLeave };
-    var dots = Array.apply(
+    let dots = Array.apply(
       null,
       Array(dotCount + 1)
         .join("0")
         .split("")
     ).map((x, i) => {
-      var leftBound = i * this.props.slidesToScroll;
-      var rightBound =
+      let leftBound = i * this.props.slidesToScroll;
+      let rightBound =
         i * this.props.slidesToScroll + (this.props.slidesToScroll - 1);
-      var isSlideActive =
+      let isSlideActive =
         this.props.currentSlide >= leftBound &&
         this.props.currentSlide <= rightBound;
+      let className = classnames({
+        "slick-active": isSlideActive,
+      });
 
-      var className = classnames({ "slick-active": isSlideActive });
-
-      var dotOptions = {
+      let dotOptions = {
         message: "dots",
         index: i,
         slidesToScroll: this.props.slidesToScroll,
-        currentSlide: this.props.currentSlide
+        currentSlide: this.props.currentSlide,
       };
 
-      var onClick = this.clickHandler.bind(this, dotOptions);
+      let onClick = this.clickHandler.bind(this, dotOptions);
       return (
         <li
           key={i}
@@ -88,14 +89,14 @@ export class Dots extends React.PureComponent {
             slidesToScroll: this.props.slidesToScroll,
             slidesToShow: this.props.slidesToShow,
             infinite: this.props.infinite,
-            name: this.props.name
+            name: this.props.name,
           })}
           aria-controls={getNthDotAssociatedSlides(i, {
             slideCount: this.props.slideCount,
             slidesToScroll: this.props.slidesToScroll,
             slidesToShow: this.props.slidesToShow,
             infinite: this.props.infinite,
-            name: this.props.name
+            name: this.props.name,
           })}
           aria-selected={isSlideActive}
           className={className}
@@ -107,8 +108,8 @@ export class Dots extends React.PureComponent {
               slidesToScroll: this.props.slidesToScroll,
               slidesToShow: this.props.slidesToShow,
               infinite: this.props.infinite,
-              name: this.props.name
-            })
+              name: this.props.name,
+            }),
           })}
         </li>
       );
